completion: recognize -c/-C when completing for git switch
If a user wants to create a new branch using git switch, we should be
able to complete any starting point reference. Additionally, DWIM logic
will not work.

Add -c and -C to the list of options for disabling --dwim and for
choosing the --refs mode.

This fixes several known breakages, so update these tests accordingly.

Signed-off-by: Jacob Keller <jacob.keller@gmail.com>
Signed-off-by: Junio C Hamano <gitster@pobox.com>